LayoutTests/platform/ios-simulator/editing/deleting/password-delete-contents-expected.txt

This makes sure we are able to delete the contents of a password field: deleting a selection and evaluating the field contents.

FAIL passwordField.value='helllo'; passwordField.setSelectionRange(3, 4); testRunner.execCommand('Delete', false, null); passwordField.value should be hello. Was helllo.
PASS successfullyParsed is true

TEST COMPLETE
